requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.7
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2
django-bootstrap5==23.1
django-crispy-forms==2.0
django-import-export==3.2.0
django-jazzmin==2.6.0
django-phonenumber-field==7.0.2
et-xmlfile==1.1.0
freetype-py==2.3.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.2
MarkupPy==1.14
MarkupSafe==2.1.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pdfkit==1.0.0
phonenumbers==8.13.10
Pillow==9.5.0
pycairo==1.23.0
python-docx==0.8.11
pytz==2023.3
PyYAML==6.0
reportlab==4.0.0
requests==2.29.0
rlPyCairo==0.2.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
sqlparse==0.4.4
tablib==3.4.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
xlrd==2.0.1
xlwt==1.3.0
